bookmark commit
OK... working from previous code, I had in mind that ignoring packages in a check was useful. Turns out, the original usage was wrong and should have '--skip-unused' instead of ignoring packages. So.... maybe this is all for naught. I guess I'll leave it for now, but add a note to possibly just take it out. Fuck... I don't know.